// src/app/pages/egresoconceptos/components/egresoconceptos-table/egresoconceptos.interface.ts
export interface EgresoconceptosInterface {
   idegresoconcepto?: number;
   fecha?: string;
   preciosiniva?: number;
   precioconiva?: number;
   cantidad?: number;
   subtotal?: number;
   total?: number;
   concepto_concepto?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}